% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activity.R
\name{account_activity}
\alias{account_activity}
\title{Get account activity.}
\usage{
account_activity()
}
\value{
A data frame.
}
\description{
Get account activity.
}
\examples{
\dontrun{
account_activity()
}
}
