% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/account.R
\name{alpaca_account}
\alias{alpaca_account}
\title{Account}
\usage{
alpaca_account()
}
\value{
Information list about account as returned by Alpaca's API
}
\description{
The account API serves important information related to an account, including
account status, funds available for trade, funds available for withdrawal,
and various flags relevant to an account’s ability to trade. An account maybe
be blocked for just for trades (trades_blocked flag) or for both trades and
transfers (account_blocked flag) if Alpaca identifies the account to engaging
in any suspicious activity. Also, in accordance with FINRA’s pattern day
trading rule, an account may be flagged for pattern day trading
(pattern_day_trader flag), which would inhibit an account from placing any
further day-trades.
}
