% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/positions.R
\name{alpaca_close_all_positions}
\alias{alpaca_close_all_positions}
\title{Close all positions}
\usage{
alpaca_close_all_positions()
}
\value{
Information list about closed positions as returned by Alpaca's API
}
\description{
Closes (liquidates) all of the account’s open long and short positions. A
response will be provided for each order that is attempted to be cancelled.
If an order is no longer cancelable, the server will respond with status 500
and reject the request.
}
