% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/positions.R
\name{alpaca_close_position}
\alias{alpaca_close_position}
\title{Close a position}
\usage{
alpaca_close_position(symbol)
}
\arguments{
\item{symbol}{Symbol on which to close position.}
}
\value{
Information list about closed position as returned by Alpaca's API
}
\description{
Closes (liquidates) the account’s open position for the given symbol. Works
for both long and short positions.
}
