% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orders.R
\name{alpaca_list_orders}
\alias{alpaca_list_orders}
\title{Get a list of orders}
\usage{
alpaca_list_orders(
  status = "open",
  limit = 50,
  after = NULL,
  until = NULL,
  direction = "desc",
  nested = TRUE
)
}
\arguments{
\item{status}{string Order status to be queried. open, closed or all. Defaults to open.}

\item{limit}{int The maximum number of orders in response. Defaults to 50 and max is 500.}

\item{after}{timestamp The response will include only ones submitted after this timestamp (exclusive.)}

\item{until}{timestamp The response will include only ones submitted until this timestamp (exclusive.)}

\item{direction}{string The chronological order of response based on the submission time. asc or desc. Defaults to desc.}

\item{nested}{boolean If true, the result will roll up multi-leg orders under the legs field of primary order.}
}
\value{
a list returned by alpaca
}
\description{
Retrieves a list of orders for the account, filtered by the supplied query
parameters.
}
\seealso{
https://alpaca.markets/docs/trading-on-alpaca/orders/
}
