% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orders.R
\name{alpaca_new_order}
\alias{alpaca_new_order}
\title{Request a new order}
\usage{
alpaca_new_order(
  symbol,
  qty = 1,
  side = "buy",
  type = "market",
  time_in_force = "day",
  limit_price = NULL,
  stop_price = NULL,
  extended_hours = FALSE,
  client_order_id = NULL,
  order_class = NULL,
  take_profit = NULL,
  stop_loss = NULL
)
}
\arguments{
\item{symbol}{symbol or asset ID to identify the asset to trade}

\item{qty}{number of shares to trade}

\item{side}{buy or sell}

\item{type}{market, limit, stop, or stop_limit}

\item{time_in_force}{stringRequired day, gtc, opg, cls, ioc, fok. Please see
Understand Orders for more info.}

\item{limit_price}{string<number>Required required if type is limit or
stop_limit}

\item{stop_price}{string<number>Required required if type is stop or
stop_limit}

\item{extended_hours}{boolean (default) false. If true, order will be
eligible to execute in premarket/afterhours. Only works with type limit and
time_in_force day.}

\item{client_order_id}{string (<= 48 characters) A unique identifier for the
order. Automatically generated if not sent.}

\item{order_class}{string simple, bracket, oco or oto. For details of
non-simple order classes, please see Bracket Order Overview}

\item{take_profit}{object Additional parameters for take-profit leg of
advanced orders limit_price string<number>Required required for bracket
orders}

\item{stop_loss}{Additional parameters for stop-loss leg of advanced orders}
}
\value{
a list returned by alpaca
}
\description{
Places a new order for the given account. An order request may be rejected if
the account is not authorized for trading, or if the tradable balance is
insufficient to fill the order.
}
\seealso{
https://alpaca.markets/docs/trading-on-alpaca/orders/
}
