% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orders.R
\name{alpaca_orders}
\alias{alpaca_orders}
\title{Orders}
\usage{
alpaca_orders(
  symbol,
  qty = 1,
  side = "buy",
  type = "market",
  time_in_force = "day",
  limit_price = NULL,
  stop_price = NULL,
  extended_hours = FALSE,
  client_order_id = NULL,
  take_profit = NULL,
  stop_loss = NULL
)
}
\arguments{
\item{symbol}{symbol or asset ID to identify the asset to trade}

\item{qty}{number of shares to trade}

\item{side}{buy or sell}

\item{type}{market, limit, stop, or stop_limit}

\item{time_in_force}{stringRequired day, gtc, opg, cls, ioc, fok. Please see
Understand Orders for more info.}

\item{limit_price}{string<number>Required required if type is limit or
stop_limit}

\item{stop_price}{string<number>Required required if type is stop or
stop_limit}

\item{extended_hours}{boolean (default) false. If true, order will be
eligible to execute in premarket/afterhours. Only works with type limit and
time_in_force day.}

\item{client_order_id}{string (<= 48 characters) A unique identifier for the
order. Automatically generated if not sent.}

\item{take_profit}{object Additional parameters for take-profit leg of
advanced orders limit_price string<number>Required required for bracket
orders}

\item{stop_loss}{Additional parameters for stop-loss leg of advanced orders}

\item{order_class}{string simple, bracket, oco or oto. For details of
non-simple order classes, please see Bracket Order Overview}
}
\value{
a list returned by alpaca
}
\description{
The Orders API allows a user to monitor, place and cancel their orders with
Alpaca. Each order has a unique identifier provided by the client. This
client-side unique order ID will be automatically generated by the system if
not provided by the client, and will be returned as part of the order object
along with the rest of the fields described below. Once an order is placed,
it can be queried using the client-side order ID to check the status. Updates
on open orders at Alpaca will also be sent over the streaming interface,
which is the recommended method of maintaining order state.
}
\seealso{
https://alpaca.markets/docs/trading-on-alpaca/orders/
}
