% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/positions.R
\name{alpaca_positions}
\alias{alpaca_positions}
\title{Positions}
\usage{
alpaca_positions()
}
\value{
Information list about account positions as returned by Alpaca's API
}
\description{
The positions API provides information about an account’s current open
positions. The response will include information such as cost basis, shares
traded, and market value, which will be updated live as price information is
updated. Once a position is closed, it will no longer be queryable through
this API.
}
