% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alpaca.R
\name{alpaca_trade_api}
\alias{alpaca_trade_api}
\title{alpaca-trade-api}
\usage{
alpaca_trade_api(env = "r-alpacar", configure = TRUE)
}
\arguments{
\item{env}{The name of the environment that contains the python
installation for alpacar to use. If not "r-alpacar," the default, this
environment will automatically be configured with
\code{\link[=configure_env]{configure_env()}}. Disable this setting with 'configure' = FALSE.}

\item{configure}{(default TRUE) If 'env' is not "r-alpacar," should the
specified environment be configured for alpacar?}
}
\value{
the alpaca-trade-api python module or NULL
}
\description{
Exposes the alpaca-trade-api python module to R using the
\href{https://rstudio.github.io/reticulate/}{reticulate} package
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
# store the module inside an R object
alpaca <- alpaca_trade_api()

# modules are read as lists with submodules nested inside
REST <- alpaca$rest$REST
TimeFrame <- alpaca$rest$TimeFrame

# module methods are accessible as callable R functions
r <- api$REST()
r$get_account()

# module attributes are accessible as R objects
api <- REST()
data_res <- api$get_bars(
  "AAPL",
  TimeFrame$Hour,
  "2021-06-08",
  "2021-06-08",
  adjustment = "raw"
)
head(data_res$df)
\dontshow{\}) # examplesIf}
}
\seealso{
\href{https://pypi.org/project/alpaca-trade-api/}{alpaca-trade-api}
for full documentation
}
