% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{alpacaDelete}
\alias{alpacaDelete}
\title{delete request to alpaca trading api (base url) used to cancel orders}
\usage{
alpacaDelete(endpoint)
}
\description{
delete request to alpaca trading api (base url) used to cancel orders
}
