% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{alpacaGet}
\alias{alpacaGet}
\title{get request to alpaca trading api (base url)}
\usage{
alpacaGet(endpoint)
}
\arguments{
\item{endpoint}{the api endpoint}
}
\description{
get request to alpaca trading api (base url)
}
