% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alpacaR.R
\docType{package}
\name{alpacaR}
\alias{alpacaR}
\title{alpacaR a wrapper for the alpaca trading api}
\description{
##What is Alpaca?
}
\details{
Alpaca is an online brokerage with a published, official REST API which is targeting the
algorithmic trader. Alpaca API lets you build and trade with real-time market data for free.
In their words they provide a portal for you to Hack Financial Systems.
\itemize{
\item \href{https://alpaca.markets/}{Introduction}
\item \href{https://app.alpaca.markets/signup}{Signup}
\item \href{https://docs.alpaca.markets/api-documentation/web-api/}{Api Documentation}
}

This is functional wrapper which returns data as tibbles, with a few exceptions.
See the package documentation for call signitures, parameter descriptions.
\subsection{Security Tokens}{

During registration at alpaca.markets you were given security tokens required to access this API.
This wrapper will read those tokens from the .env file. The .env file must include the
following three key pairs: note(put .env in your .gitignore file)\preformatted{APCA-API-KEY-ID = "..."
APCA-API-SECRET-KEY = "..."
}
}
}
\author{
Bruce C. Schardt, \email{schardt.bruce.curtis@gmail.com}
}
