% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assets.R
\name{assets_get}
\alias{assets_get}
\title{Get information on a specific asset.}
\usage{
assets_get(id)
}
\arguments{
\item{id}{A symbol or asset ID.}
}
\value{
A data frame.
}
\description{
Get information on a specific asset.
}
\examples{
\dontrun{
AAPL <- assets_get("AAPL")
AMD <- assets_get("03fb07bb-5db1-4077-8dea-5d711b272625")
}
}
