% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/authenticate.R
\name{authenticate}
\alias{authenticate}
\title{Set Alpaca REST API key & secret}
\usage{
authenticate(key = NULL, secret = NULL)
}
\arguments{
\item{key}{API key.}

\item{secret}{API secret.}
}
\description{
Set Alpaca REST API key & secret
}
\examples{
authenticate(
  key = Sys.getenv("ALPACA_KEY"),
  secret = Sys.getenv("ALPACA_SECRET_KEY")
)
}
