% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stocks.R
\name{bars}
\alias{bars}
\title{Get historic bars}
\usage{
bars(symbols, timeframe, start, end)
}
\arguments{
\item{symbols}{One or more symbols.}
}
\value{
A data frame.
}
\description{
Get historic bars
}
\examples{
\dontrun{
}
}
