% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/http.R
\name{base_url}
\alias{base_url}
\title{Set or query API base URL}
\usage{
base_url(url = "https://api.alpaca.markets")
}
\arguments{
\item{url}{Base URL (with or without trailling slash).}
}
\value{
URL string.
}
\description{
Set or query API base URL
}
\examples{
base_url()
# The live API.
base_url("https://api.alpaca.markets")
# The paper-trading API.
base_url("https://paper-api.alpaca.markets/v2")
}
