% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trading.R
\name{calendar}
\alias{calendar}
\title{market calendar as dataframe}
\usage{
calendar(start = NULL, end = NULL)
}
\arguments{
\item{start}{date | NULL}

\item{end}{date | NULL}
}
\description{
market calendar as dataframe
}
