% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clock.R
\name{clock}
\alias{clock}
\title{Get market clock information.}
\usage{
clock()
}
\value{
A list.
}
\description{
Returns the following information:
}
\details{
- the current market timestamp (`timestamp`),
- whether or not the market is currently open (`is_open`), and
- the times of the next market open (`next_open`) and close (`next_close`).
}
\examples{
\dontrun{
clock()
}
}
