% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stocks.R
\name{condition_codes}
\alias{condition_codes}
\title{Get condition codes.}
\usage{
condition_codes(tape, ticktype = "quote")
}
\arguments{
\item{symbols}{One or more symbols.}
}
\value{
A data frame.
}
\description{
Get condition codes.
}
\examples{
\dontrun{
condition_codes("C")
}
}
