% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/environment.R
\name{configure_env}
\alias{configure_env}
\title{Configure python environment for alpacar}
\usage{
configure_env(conda_env = "r-alpacar")
}
\arguments{
\item{conda_env}{(optional) The name of the environment to be created or configured for
alpacar If not "r-alpacar," the default, this environment must be
specified in \code{\link[=alpaca_trade_api]{alpaca_trade_api()}}.}
}
\value{
\code{\link[=invisible]{invisible()}}
}
\description{
Creates a conda environment if needed, then configures that environment with
the correct modules and python version for the alpacar package.
}
\section{Note}{

Requires miniconda installation: https://docs.conda.io/en/latest/miniconda.html
}

