% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stocks.R
\name{exchange_codes}
\alias{exchange_codes}
\title{Get exchange codes.}
\usage{
exchange_codes()
}
\arguments{
\item{symbols}{One or more symbols.}
}
\value{
A data frame.
}
\description{
Get exchange codes.
}
\examples{
\dontrun{
exchange_codes()
}
}
