% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trading.R
\name{newTrade}
\alias{newTrade}
\title{create a new Trade object}
\usage{
newTrade(
  symbol,
  side,
  qty,
  time_in_force = "day",
  limit_price = NULL,
  stop_price = NULL,
  id = NULL
)
}
\arguments{
\item{symbol}{a market symbol}

\item{side}{"buy" | "sell"}

\item{qty}{number of shares}

\item{time_in_force}{"day" | "gtc" | "opg"}

\item{limit_price}{NULL | limit price for limt and sotp limit type orders}

\item{stop_price}{NULL | stop price for stop and stop_limit type orders}
}
\description{
create a new Trade object
}
