% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orders.R
\name{order_create}
\alias{order_create}
\title{Place an order.}
\usage{
order_create(symbol, side, quantity, type = "market", limit_price = NA)
}
\value{
A data frame.
}
\description{
Place an order.
}
\examples{
\dontrun{
order_create("SPY", "buy", 1, "limit")
}
}
