% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/account.R
\name{print.alpacaAccount}
\alias{print.alpacaAccount}
\title{Generic print() method for Alpaca Account objects.}
\usage{
\method{print}{alpacaAccount}(obj)
}
\arguments{
\item{obj}{An Alpaca Account object.}
}
\description{
Generic print() method for Alpaca Account objects.
}
\examples{
\dontrun{
print(account())
account()
}
}
