% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-tbl.R
\name{tbl}
\alias{tbl}
\title{tbl}
\usage{
tbl(
  ...,
  .rows = NULL,
  .name_repair = c("check_unique", "unique", "universal", "minimal")
)
}
\description{
See \code{tibble::\link[tibble:tibble]{tibble}} for details.
}
\keyword{internal}
