% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stocks.R
\name{trades}
\alias{trades}
\title{Get trades.}
\usage{
trades(symbols, start, end, feed = "iex")
}
\arguments{
\item{symbols}{One or more symbols.}

\item{start}{Start date (YYYY-MM-DD).}

\item{end}{End date (YYYY-MM-DD).}

\item{feed}{Data source. Either `"sip"`, `"iex"` (default) or `"otc"`.}
}
\value{
A data frame.
}
\description{
Get trades.
}
\examples{
\dontrun{
}
}
