% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/watchlist.R
\name{watchlist}
\alias{watchlist}
\title{Get watchlist by ID.}
\usage{
watchlist(id)
}
\arguments{
\item{id}{Watchlist ID.

The `assets` column is a list column containing a data frame listing the watched assets.}
}
\value{
A data frame.
}
\description{
Get watchlist by ID.
}
\examples{
\dontrun{
}
}
