% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/watchlist.R
\name{watchlist_add}
\alias{watchlist_add}
\title{Add asset to watchlist by ID.}
\usage{
watchlist_add(id, symbol)
}
\arguments{
\item{id}{Watchlist ID.}

\item{symbol}{Asset symbol.}
}
\value{
A data frame.
}
\description{
Add asset to watchlist by ID.
}
\examples{
\dontrun{
}
}
