% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{alpacaPost}
\alias{alpacaPost}
\title{post request to alpaca trading api (base url) used to submit orders}
\usage{
alpacaPost(endpoint, body)
}
\description{
post request to alpaca trading api (base url) used to submit orders
}
