% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{options_list}
\alias{options_list}
\title{Get options data.}
\usage{
options_list(
  underlying_symbols,
  type = NULL,
  expiration_date_gte = NULL,
  expiration_date_lte = NULL
)
}
\arguments{
\item{type}{Filter contracts by the type (`"call"` or `"put"`).}

\item{expiration_date_gte}{Minimum expiration date.}

\item{expiration_date_lte}{Maximum expiration date.}
}
\value{
A data frame.
}
\description{
Get options data.
}
\examples{
\dontrun{
options("AAPL")
options("AAPL", "put")
}
}
