% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stocks.R
\name{quotes_latest}
\alias{quotes_latest}
\title{Get latest quotes for one or more stocks.}
\usage{
quotes_latest(symbols)
}
\arguments{
\item{symbols}{One or more symbols.}
}
\value{
A data frame.
}
\description{
Get latest quotes for one or more stocks.
}
\examples{
\dontrun{
quotes_latest("AAPL")
quotes_latest(c("AAPL", "TSLA"))
}
}
